Changed Visitor.borrows to Visitors.borrowing to show it only keeps track of borrows that aren't complete.

+++ src/VisitorRegistry/Visit.java
@@ -5,13 +5,13 @@ import java.time.*;
  * Stores data on a single visit.
  */
 public class Visit {
-    private int visitorID;
+    private long visitorID;
     private LocalDate dateOfVisit;
     private LocalTime timeOfArrival;
     private LocalTime timeOfDeparture;
     private boolean ongoing;
 
-    public Visit(int visId, LocalDate visitDate, LocalTime time) {
+    public Visit(long visId, LocalDate visitDate, LocalTime time) {
         visitorID = visId;
         dateOfVisit = visitDate;
         timeOfArrival = time;
